Move NaslValue into naslvalue.rs
Instead of having NaslValue defined in interpreter.rs it is moved into
naslvalue.rs
