+++ rust/nasl-interpreter/src/built_in_functions/description.rs
@@ -7,7 +7,7 @@ use std::str::FromStr;
 use crate::{
     context::{ContextType, Register},
     error::FunctionError,
-    interpreter::NaslValue,
+    NaslValue,
     NaslFunction,
 };
